requirements.txt
streamlit
snowflake-connector-python
pyyaml
openai>=1.0.0
graphviz
